+++ app/src/main/java/hexlet/code/repository/UrlRepository.java
@@ -17,14 +17,14 @@ public class UrlRepository extends BaseRepository {
             preparedStatement.executeUpdate();
             var generatedKeys = preparedStatement.getGeneratedKeys();
             if (generatedKeys.next()) {
-                url.setId(generatedKeys.getInt(1));
+                url.setId(generatedKeys.getLong(1));
             } else {
                 throw new SQLException("DB have not returned an id after saving an entity");
             }
         }
     }
 
-    public static Optional<Url> find(Integer id) throws SQLException {
+    public static Optional<Url> find(Long id) throws SQLException {
         var sql = "SELECT * FROM urls WHERE id = ?";
         try (var conn = dataSource.getConnection();
              var stmt = conn.prepareStatement(sql)) {
@@ -49,7 +49,7 @@ public class UrlRepository extends BaseRepository {
             stmt.setString(1, urlName);
             var resultSet = stmt.executeQuery();
             if (resultSet.next()) {
-                var id = resultSet.getInt("id");
+                var id = resultSet.getLong("id");
                 var name = resultSet.getString("name");
                 var createdAt = resultSet.getTimestamp("created_at");
                 var url = new Url(name);
@@ -68,7 +68,7 @@ public class UrlRepository extends BaseRepository {
             var resultSet = stmt.executeQuery();
             var result = new ArrayList<Url>();
             while (resultSet.next()) {
-                var id = resultSet.getInt("id");
+                var id = resultSet.getLong("id");
                 var name = resultSet.getString("name");
                 var createdAt = resultSet.getTimestamp("created_at");
                 var url = new Url(name);
